requirements.txt
livelossplot==0.5.4
opencv-python-headless==4.5.3.56
matplotlib==3.4.3
numpy==1.21.2
pillow==7.0.0
bokeh==2.1.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.63.0
ipywidgets==7.6.5
pytest==7.1.1
pandas==1.3.5
seaborn==0.11.2
streamlit
viola